requirements.txt
Django==2.0
psycopg2
freeze
djangorestframework


django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
